seperate option menu & title scene
